bugfix for ubuntu

+++ distiller/include/DistillerConstants.hh
@@ -1,12 +1,14 @@
 #ifndef DISTILLER_CONSTANTS_HH
 #define DISTILLER_CONSTANTS_HH
 
+#include <cstddef>
+
 const static float MIN_DPHI_JET_MET = 0.4; 
 const static float REMOVE_JET_CONE = 0.2; 
 const static float REMOVE_EL_CONE = 0.4; 
 const static float REMOVE_MU_CONE = 0.4; 
 
-const static int N_JETS_TO_SAVE = 4; 
+const static size_t N_JETS_TO_SAVE = 4; 
 const static size_t N_SR_JETS = 3; 
 
 const static float VETO_ELECTRON_REL_ISOLATION = 0.1;
